% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_comps.R
\name{aggregate_comps}
\alias{aggregate_comps}
\title{Function to aggregate compositions into a shorter range of bins based on  data availability in each bin}
\usage{
aggregate_comps(comp.data, compbins = NULL, compression = 1,
  proportion.thresh = 0.9)
}
\arguments{
\item{comp.data}{output from the \code{calc_age2length} function, namely a list with objects mulen, muweight, and natlength}

\item{compbins}{the age or length composition bins you want to compress the data into, either this or compression can be specified but not both}

\item{compression}{the minimum thresshold for numbers of fish in each bin for the bin to be included in the data, default is 1}
}
\description{
Function to aggregate compositions into a shorter range of bins based on  data availability in each bin
}
