% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{aggregate and select data from an Atlantis scenario}
\usage{
aggregateData(dat, time, species, boxes, keepColumns = c("species", "agecl",
  "polygon"))
}
\arguments{
\item{dat}{The dataframe of numbers-at-age from reading in the Atlantis fileSnapshot}

\item{time}{The timing of the survey (can be a vector if survey spans multiple time periods)}

\item{species}{The species to sample in the survey (a vector)}

\item{boxes}{A vector of box numbers}
}
\description{
aggregate and select data from an Atlantis scenario
}
\details{
Returns a subsetted matrix with columns not aggregated over

columns: species, agecl, polygon, layer, time, atoutput

--will sum over layers, but enter NA as layer to indicate all layers

This function is for a vector of defined species
}
\examples{
dat <- data.frame(species = c(rep("spec1",3*3),rep("spec2",5*3)),
  agecl = c(rep(1:3,3),rep(3:7,3)),
  polygon = c(rep(1:3,each=3),rep(1:3,each=5)),
 layer = 1, time = 1)
dat$atoutput <- 10000/dat$agecl
boxes <- 1:2
tmp <- aggregateData(dat=dat, time=1, species=c("spec1","spec2"), boxes = boxes)
}
\author{
Poseidon
}

