% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age2length.R
\name{calc_age2length}
\alias{calc_age2length}
\title{Calculate length from ages}
\usage{
calc_age2length(
  structn,
  resn,
  nums,
  biolprm,
  fgs,
  maxbin = 150,
  CVlenage = 0.1,
  remove.zeroes = TRUE
)
}
\arguments{
\item{structn}{A \code{data.frame} containing the structural nitrogen per functional-group
("species") per timestep, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "StructN"}.}

\item{resn}{A \code{data.frame} containing the reserve nitrogen per functional-group
("species") per timestep, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "ResN"}.}

\item{nums}{A \code{data.frame} containing the numbers per functional-group
("species") per timesetp, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "Nums"}.}

\item{biolprm}{A list of biological parameters available from the
\code{[...]_Biol.prm} file, as read in by \code{\link{load_biolprm}}.}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{maxbin}{The upper length (cm) bin applied to all species sampled. The default
value is 150.
Optionally, a \code{data.frame} can be supplied for multiple species with different
species lengths: Columns are:
                species: the species name. Matches names in species
                maxlenbin:  the upper length (cm) bin applied to that species}

\item{CVlenage}{The variability in length at age (same for all species)
The default value is 0.1.
Optionally, a \code{data.frame} can be supplied for multiple species with different
CV of length at age: Columns are:
                species: the species name. Matches names in species
                cvlenage:  the variability in length at age for that species}

\item{remove.zeroes}{Logical asking whether to only return numbers at length for
combinations of species, age, box, depth, etc that have numbers>0.
The default value is \code{TRUE}.}
}
\value{
A \code{list} containing three \code{data.frame}s, mulen (mean length at age),
muweight (mean weight at age), and natlength (numbers at length). natlength is in the
same format as other dataframes in the atlantisom package except has two additional
columns to include the length bin information.
}
\description{
Calculate length composition in 1 cm bins from Atlantis output (st)age data
Uses numbers at age, plus weight at age, and weight-length relationships to
generate size comps
}
\examples{
\dontrun{
directory <- system.file("extdata", "SETAS_Example", package = "atlantisom")
load(file.path(directory, "outputsrun_truth.RData"))
# WARNING: EXTREMELY LONG RUN TIME for all species at full model resolution!
# even one species has a long run time; this works but takes minutes to return
result$structn <- result$structn[result$structn$species == "Demersal_S_Fish",]
result$resn <- result$resn[result$resn$species == "Demersal_S_Fish",]
result$nums <- result$nums[result$nums$species == "Demersal_S_Fish",]
lenout <- calc_age2length(
  structn=result$structn, resn=result$resn, nums=result$nums,
  biolprm = result$biolprm, fgs = result$fgs,
  CVlenage = 0.1, remove.zeroes=TRUE)
str(lenout)
rm(lenout)
}
}
\seealso{
Other calc functions: 
\code{\link{calc_Z}()},
\code{\link{calc_avgwtstage2age}()},
\code{\link{calc_biomass_age}()},
\code{\link{calc_biomass_pool}()},
\code{\link{calc_pred_cons}()},
\code{\link{calc_pred_diet}()},
\code{\link{calc_stage2age}()},
\code{\link{calc_timestep2time}()}
}
\author{
Gavin Fay
}
\concept{calc functions}
