% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_avgwtstage2age.R
\name{calc_avgwtstage2age}
\alias{calc_avgwtstage2age}
\title{Approximating the average weight at true age from stage-specific average weight.}
\usage{
calc_avgwtstage2age(wtagecl, annages, fgs)
}
\arguments{
\item{wtagecl}{average weight at age output of the \code{calc_age2length} function}

\item{annages}{truenumsage output of \code{run_truth}}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}
}
\value{
A \code{data.frame} in long format with the following column names:
  species, time, polygon, agecl, and atoutput (i.e., variable). This is
  the same as the input Nums data frame however, agecl will be TRUE numbers
  at annual age -- just named the same as the input to ensure make it work
  with the other functions
}
\description{
Approximating the average weight at true age from stage-specific average weight.
}
\examples{
dir <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
file_nc="outputSETAS.nc"
fgs=load_fgs(dir = dir, "functionalGroups.csv")
file_init="INIT_VMPA_Jan2015.nc"
bps=load_bps(dir = dir, fgs, file_init)
select_groups=fgs$Name[fgs$IsTurnedOn > 0]
select_variable="Nums"
box.info=load_box(dir = dir, file_bgm="VMPA_setas.bgm")
bboxes=get_boundary(box.info)
#when calc_stage2age is run in the run_truth, it will need to have the nums
#data frame and the bioprm already read in:
nums_data <- load_nc(dir = dir,
                     file_nc="outputSETAS.nc",
                     bps=bps, fgs=fgs, select_groups=select_groups,
                     select_variable = "Nums",
                     check_acronyms = TRUE, bboxes = bboxes)
biolprm <- load_biolprm(dir, file_biolprm="VMPA_setas_biol_fishing_Trunk.prm")
YOY <- load_yoy(dir, file_yoy="outputSETASYOY.txt")
}
\seealso{
Other calc functions: 
\code{\link{calc_Z}()},
\code{\link{calc_biomass_age}()},
\code{\link{calc_pred_diet}()},
\code{\link{calc_stage2age}()},
\code{\link{calc_timestep2time}()}
}
\author{
Sarah Gaichas
}
\concept{calc functions}
