% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biomass_age.R
\name{calc_biomass_age}
\alias{calc_biomass_age}
\title{Calculate biomass in t per age-structured-groups per polygon, time and ageclass.}
\usage{
calc_biomass_age(nums, resn, structn, biolprm)
}
\arguments{
\item{nums}{A \code{data.frame} containing the numbers per functional-group
("species") per timesetp, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "Nums"}.}

\item{resn}{A \code{data.frame} containing the reserve nitrogen per functional-group
("species") per timestep, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "ResN"}.}

\item{structn}{A \code{data.frame} containing the structural nitrogen per functional-group
("species") per timestep, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "StructN"}.}

\item{biolprm}{A list of biological parameters available from the
\code{[...]_Biol.prm} file, as read in by \code{\link{load_biolprm}}.}
}
\value{
A \code{data.frame} in tidy format with the following columns:
  species, agecl, time, polygon and atoutput.
  Atoutput is the total biomass in t.
  At this point biomass was aggregated (sum) per layer!
}
\description{
Calculate biomass in t per age-structured-groups per polygon, time and ageclass.
}
\details{
This functions converts the ATLANTIS output to a \code{data.frame}
  which can be processed in R.
}
\author{
Alexander Keth
}
\seealso{
Other calc functions: \code{\link{calc_Z}},
  \code{\link{calc_stage2age}}, \code{\link{convert_time}}
}
\keyword{gen}

