% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pred_diet.R
\name{calc_pred_diet}
\alias{calc_pred_diet}
\title{Calculate eaten biomass in t for each functional group.}
\usage{
calc_pred_diet(dietcomp, eat, grazing, vol, biolprm)
}
\arguments{
\item{dietcomp}{Dataframe containing the diet proportion for each functional group
per prey, time and ageclass. The dataframe has to origin from load_diet_comp using the
diet_check.txt as input.}

\item{eat}{Dataframe containing the consumed biomass in mg N for each functional-group ("species")
per timesetp, ageclass, and polygon. The dataframe has to origin from load_nc using "Eat" as select variable.}

\item{grazing}{Dataframe containing the consumed biomass in mg N for each functional-group ("species")
per timesetp, ageclass, and polygon. The dataframe has to origin from load_nc using "Grazing" as select variable.}

\item{biolprm}{A list of biological parameters available from the
\code{[...]_Biol.prm} file, as read in by \code{\link{load_biolprm}}.}
}
\value{
Dataframe in tidy format with the following columns: species, agecl, polygon, time,
atoutput, vol, dietcomp, prey, bio_eaten.
}
\description{
Function to calculate the eaten biomass in t for each functional group per
time, polygon, ageclass and preyitem. Please note, that the function
produces unrealistic high consumption values at the momemt.
}
\author{
Alexander Keth
}

