% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stage2age.R
\name{calc_stage2age}
\alias{calc_stage2age}
\title{Approximating the true numbers at age within each box, functional group and
time step.}
\usage{
calc_stage2age(dir, nums_data, biolprm, YOY_name)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}
}
\value{
A \code{data.frame} in long format with the following coumn names:
  Species, timestep, polygon, TRUEagecl, and atoutput (i.e., variable).
}
\description{
Will be called in the run_atlantis file, and requires inputs from
\code{\link{load_nc}} for the "Nums" variable, and the \code{\link{load_meta}}
and finally the \code{\link{load_biolprm}}.
}
\examples{
#This is just to bring in an example data frame I can use to write the code,
#much of the beginning code will be removed as I actually write the function,
#but I wanted to push what I have so far!
dir <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
file_nc="outputSETAS.nc"
fgs=load_fgs(dir = dir, "functionalGroups.csv")
file_init="INIT_VMPA_Jan2015.nc"
bps=load_bps(dir = dir, fgs, file_init)
select_groups=fgs$Name[fgs$IsTurnedOn > 0]
select_variable="Nums"
box.info=load_box(dir = dir, file_bgm="VMPA_setas.bgm")
bboxes=get_boundary(box.info)
#when calc_stage2age is run in the run_atlantis, it will need to have the nums
#data frame and the bioprm already read in:
nums_data <- load_nc(dir = dir,
                     file_nc="outputSETAS.nc",
                     bps=bps, fgs=fgs, select_groups=select_groups,
                     select_variable = "Nums",
                     check_acronyms = TRUE, bboxes = bboxes)
biolprm <- load_biolprm(dir, file_biolprm="VMPA_setas_biol_fishing_Trunk.prm")
recruits <- as.data.table(read.table(YOY, header = T))
}
\author{
Emma Hodgson
}
\seealso{
Other calc functions: \code{\link{calc_Z}},
  \code{\link{calc_biomass_age}},
  \code{\link{convert_time}}
}

