% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_timestep2time.R
\name{calc_timestep2time}
\alias{calc_timestep2time}
\title{Converts timestep from Atlantis output to real time}
\usage{
calc_timestep2time(dir = getwd(), data, file_runprm)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{data}{A data frame of Atlantis output. Where the data frame can be
of any structure as long as it has a column with the name \code{"time"}.}

\item{file_runprm}{A character value giving the file name of the biology
\code{.xml} file.
The file should be located in your current working directory or the
folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.
Usually the file is named \code{"[...]_run.xml".}.
This file is output from an Atlantis run in a standard format and is based
on input from the \code{"[...]_run.prm"} file.}
}
\value{
A \code{data.frame} in tidy format with a column date showing
  the real date and year showing the real year.
  If there is not column \code{"time"} in the \code{data.frame} specified
  for the \code{data}, then the original \code{data.frame}
  is returned.
}
\description{
Converts timestep from Atlantis output to real time
}
\details{
Atlantis output uses time steps, where time steps are sequential
  and start at zero. \code{convert_time} converts the time step to dates,
  where zero is January 01, 1970.
}
\examples{

d <- system.file("extdata", "SETAS_Example", package = "atlantisom")
load(file = file.path(d, "outputsrun_truth.RData"))
test <- calc_timestep2time(dir = d, data = result$biomass_ages,
  file_runprm = "Run_settings.xml")
rm(test)

}
\seealso{
\code{\link{load_fgs}}

Other calc functions: 
\code{\link{calc_Z}()},
\code{\link{calc_age2length}()},
\code{\link{calc_avgwtstage2age}()},
\code{\link{calc_biomass_age}()},
\code{\link{calc_pred_cons}()},
\code{\link{calc_pred_diet}()},
\code{\link{calc_stage2age}()}
}
\author{
Alexander Keth
}
\concept{calc functions}
