% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_timestep2time.R
\name{convert_time}
\alias{convert_time}
\title{Converts timestep from Atlantis output to actual time!}
\usage{
convert_time(dir, data, file_runprm, modelstart)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{data}{todo: document this properly}

\item{file_runprm}{A character value giving the file name of the biology
\code{.prm} file.
The file should be located in your current working directory or the
folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.
Usually the file is named \code{"[...]_run.prm".}.}

\item{modelstart}{todo: document this properly}
}
\value{
dataframe in tidy format with column time showing the actual time.
}
\description{
Converts timestep from Atlantis output to actual time!
}
\examples{
testscenario <- "INIT_VMPA_Jan2015"
d <- system.file("extdata", testscenario, package = "atlantisom")
load(file = file.path(d, "outputSETASrun_atlantis.RData"))
data <- result$biomass_ages
#runfile <- ".prm"
#test <- convert_time(dir = d, data = data, file_runprm = runfile)
}
\author{
Alexander Keth
}
\seealso{
\code{\link{load_fgs}}

Other calc functions: \code{\link{calc_Z}},
  \code{\link{calc_biomass_age}},
  \code{\link{calc_stage2age}}
}

