% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env.R
\name{create_env}
\alias{create_env}
\title{A function to sample from the environmental truth provided
by the Atlantis scenario.}
\usage{
create_env(true_env, subset)
}
\arguments{
\item{true_env}{A \code{data.frame} of true environmental data
sampled from an Atlantis scenario.}

\item{subset}{A \code{data.frame} specifying which samples to keep}
}
\description{
Uses data provided by an Atlantis scenario and adds
  observation error and/or bias, along with subsetting, to provide
  an environmental index for use in subsequent model fitting.
}
\author{
Kelli Faye Johnson
}
\concept{create functions}
