% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fishery_subset.R
\name{create_fishery_data}
\alias{create_fishery_data}
\title{Create fishery data from an Altantis scenario}
\usage{
create_fishery_data(dat, time, species, boxes)
}
\arguments{
\item{dat}{A \code{data.frame} of numbers at age containing the following columns:
\itemize{
  \item{species}
  \item{agecl}
  \item{polygon}
  \item{layer}
  \item{time}
  \item{atoutput}
}
The \code{data.frame} is generated from either \code{\link{create_survey}}
or \code{\link{create_fishery_data}}.}

\item{time}{The timing of the survey (a vector indicating specific time steps, which are typically associated with years)
i.e., seq(365,10*3650,365) would be an annual survey for 10 years}

\item{species}{The species to sample in the survey (a vector)}

\item{boxes}{A vector of box numbers}

\item{effic}{Efficiency for each species: a matrix with nrow=length(species). Columns:
 species:    the species name. Matches names in species
efficiency:}

\item{selex}{Selectivity at age. A dataframe defining selectivity at age for each species. Columns are:
species: the species name. Matches names in species
agecl:   the age class that selectivity represents
selex:   the proportion selected relative to fully selected age classes (between 0 and 1)}
}
\value{
A \code{matrix} in the same format as the \code{dat} summed
  over layers, with \code{NA} in the layer column.
}
\description{
Create fishery data for an Atlantis scenario, where
  observation error will be added and certain boxes and species will
  be sampled.
  todo: add more information here.
}
\details{
The function works for a vector of defined species
  and only returns information in boxes that are sampled.
}
\author{
Poseidon
}

