% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey_env.R
\name{create_survey_env}
\alias{create_survey_env}
\title{Create environmental survey data subset from Atlantis output}
\usage{
create_survey_env(true_env, survdepth, survboxes, survtimes)
}
\arguments{
\item{true_env}{A \code{data.frame} of true environmental data
sampled from an Atlantis scenario.}

\item{time}{The timing of the survey (a vector indicating specific time steps, which are typically associated with years)
i.e., seq(365,10*3650,365) would be an annual survey for 10 years}

\item{layer}{Which layer to sample, surface or bottom? or numeric model layer}

\item{boxes}{A vector of box numbers}
}
\value{
The function returns a subsetted matrix with the same columns
  as the input data, i.e.,:
  \itemize{
    \item{variable}
    \item{polygon}
    \item{layer}
    \item{time}
    \item{atoutput}
  }
}
\description{
Create a subset of the physical environment observations for an Atlantis scenario.

Uses data provided by an Atlantis scenario and adds
  observation error and/or bias, along with subsetting, to provide
  an environmental index for use in subsequent model fitting.
}
\details{
The function takes physical data from an Atlantis scenario
  where the data was read in from Atlantis output using \code{\link{load_nc_physics}}.
  One does not need to use these functions to create \code{dat}, rather you must only
  ensure that the structure of \code{dat} is the same.
  Currently, the function subsets the data by polygon and time, with the user defining
  a single layer, either "bottom" or "surface", to sample. Possible environmental variables
  from \code{\link{load_nc_physics}} include "salt", "NO3", "NH3", "Temp", "Oxygen", "Si",
  "Det_Si", "DON", "Chl_a", "Denitrifiction", "Nitrification".
  Only one variable is loaded with a call to \code{\link{load_nc_physics}}, so this
  function also takes one environmental variable at a time and subsets to specific
  defined polygons, and specific defined time representing a survey.
}
\author{
Kelli Faye Johnson, Sarah Gaichas
}
\concept{create functions}
