% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundary.R
\name{get_boundary}
\alias{get_boundary}
\title{Get boundary boxes from Atlantis box information.}
\usage{
get_boundary(boxinfo)
}
\arguments{
\item{boxinfo}{A \code{list} as returned from \code{\link{load_box}}.}
}
\value{
A \code{vector} specifying which boxes are on the boundary.
}
\description{
Use the output from \code{\link{load_box}} and obtain a \code{vector}
specifying which boxes are along the boundary.
}
\examples{
d <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
boxes <- get_boundary(load_box(dir = d, file_bgm = "VMPA_setas.bgm"))
rm(boxes)

}
\seealso{
\code{\link{load_box}}
}
\author{
Kelli Faye Johnson
}
\concept{get functions}
