% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_acronyms}
\alias{get_age_acronyms}
\alias{get_age_groups}
\alias{get_fish_acronyms}
\alias{get_groups}
\alias{get_nonage_acronyms}
\title{Collection of similar functions which get specific
columns from the Atlantis \code{functionalGroups.csv}}
\usage{
get_groups(fgs)

get_age_groups(fgs)

get_acronyms(fgs)

get_age_acronyms(fgs)

get_nonage_acronyms(fgs)

get_fish_acronyms(fgs)
}
\arguments{
\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}
}
\value{
Character string.
}
\description{
This collection of functions takes loaded functional
groups, via the \code{\link{load_fgs}} function and creates various
character strings of group names or acronym names.
}
\details{
Currently, the following character strings can be created
- get_groups extracts the column "Name"
- get_age_groups extracts the column "Name". Selects groups with 10 ageclasses.
- get_acronym extracts the column "Code"
- get_age_acronym extracts the column "Code". Selects groups with 10 ageclasses.
- get_nonage_acronym extracts the columne "Code". Only groups with ageclasses different from 10 are selected.
- get_fish_acronyms ectracts the column "Code". Only groups with InvertType equal to "FISH" or "SHARK" are selected.
}
\seealso{
Other get functions: \code{\link{get_boundary}}
}

