% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_groups.R
\name{load_bps}
\alias{load_bps}
\title{This function still need to load in the initial conditions file to check the number of
dimension in the "N" variable. Models with an updated version of the functional-groups
file will be able to load the "biomasspools" (which aren't true biomasspools but all epibenthic
groups) from the functional groups file directly. This will save considerable amount
of time. In addition parameter handling will be much easier, as there is no need to pass
the init-file to the function call anymore.}
\usage{
load_bps(file_fgs, nc_init)
}
\description{
This function still need to load in the initial conditions file to check the number of
dimension in the "N" variable. Models with an updated version of the functional-groups
file will be able to load the "biomasspools" (which aren't true biomasspools but all epibenthic
groups) from the functional groups file directly. This will save considerable amount
of time. In addition parameter handling will be much easier, as there is no need to pass
the init-file to the function call anymore.
}

