% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_atlantis_ncdf.R
\name{load_atlantis_ncdf}
\alias{load_atlantis_ncdf}
\title{Load Atlantis outputfiles (netcdf)}
\usage{
load_atlantis_ncdf(nc_out, file_fgs, select_groups, select_variable,
  remove_bboxes, check_acronyms)
}
\arguments{
\item{nc_out}{Connection of the ATLANTIS output file given as complete folder/filename string. Usually "output[...].nc".
Additionally "output[...]PROD.nc" can be loaded.}

\item{file_fgs}{Connection of the ATLANTIS functional groups file given as complete folder/filename string.
Usually "functionalGroups.csv".}

\item{select_groups}{Character vector of funtional groups which shall be plotted. Names have to match the ones
used in the ncdf file. Check column "Name" in "functionalGroups.csv" for clarification.}

\item{select_variable}{Character value spefifying which variables shall be loaded. Only one variable
can be loaded at a time. Currently, only "N", "Nums", "ResN", "StructN", "Eat", "Growth", "Prodn",
"Grazing" can be selected.}

\item{remove_bboxes}{Logical specifying if an boundary boxes  shall be excluded. Default is T.}

\item{check_acronyms}{Logical specifying if selected groups are active in the model run. This is used in automated runs.
Since all groups are passed when plotting (plot_atlantis) is called via batch-file (which cannot be changed easily)
this will result in errors if some groups are not active in the model run. By default this is TRUE.}
}
\value{
Dataframe in long format with the following informations: Species, timestep, polygon, agecl and value (= "atoutput").
}
\description{
This function loads Atlantis outputfiles (netcdf) and converts them to a dataframe.
}
\details{
This functions converts the ATLANTIS output to a dataframe which can be processed in R.
}
\examples{
load_atlantis_output(model_path = file.path("z:", "Atlantis", "ATLANTIS NSmodel base"), filename = "outputNorthSea.nc", select_groups = get_groups(), select_variable = "ResN", biomasspools = c("large_crabs", "small_epifauna", "sessile_epifauna", "epifaunal_macrobenthos"))
}
\keyword{gen}

