% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_biolprm.R
\name{load_biolprm}
\alias{load_biolprm}
\title{Load Atlantis \code{_Biol.prm} file}
\usage{
load_biolprm(dir = getwd(), file_biolprm)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_biolprm}{A character value giving the file name of the biology
\code{.prm} file.
The file should be located in your current working directory or the
folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.
Usually the file is named \code{"[...]_Biol.prm".}.}
}
\value{
A list of biological parameters from the \code{_Biol.prm} file.
}
\description{
This function loads the Atlantis biology parameter output file.
}
\examples{
d <- system.file("extdata", "SETAS_Example", package = "atlantisom")
biolprm <- load_biolprm(d, "Biology.prm")
rm(biolprm)

}
\seealso{
Other load functions: 
\code{\link{load_agebioind}()},
\code{\link{load_bioind}()},
\code{\link{load_boxarea}()},
\code{\link{load_box}()},
\code{\link{load_bps}()},
\code{\link{load_catch}()},
\code{\link{load_detailed_diet_comp}()},
\code{\link{load_diet_comp}()},
\code{\link{load_fgs}()},
\code{\link{load_fisheries}()},
\code{\link{load_meta}()},
\code{\link{load_nc_annage}()},
\code{\link{load_nc_catchtons}()},
\code{\link{load_nc}()},
\code{\link{load_runprm}()},
\code{\link{load_yoy}()}
}
\author{
Kelli Faye Johnson
}
\concept{load functions}
