% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_box.R
\name{load_box}
\alias{load_box}
\title{Load the box specification file for an Atlantis scenario}
\usage{
load_box(dir = getwd(), file_bgm)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_bgm}{A character value specifying the file name of a \code{.bgm}
  file, which contains information on box locations for an Atlantis scenario.}
}
\value{
A list of information regarding boxes for an Atlantis scenario.
}
\description{
Read in the \code{.bgm} file that specifies the the box coordinates
for an Atlantis scenario.
}
\author{
Kelli Faye Johnson
}
\seealso{
Other load functions: \code{\link{load_acronyms}},
  \code{\link{load_age_acronyms}},
  \code{\link{load_age_groups}},
  \code{\link{load_fish_acronyms}},
  \code{\link{load_groups}},
  \code{\link{load_nonage_acronyms}},
  \code{\link{read_functionalgroups}};
  \code{\link{load_boxarea}}; \code{\link{load_diet_comp}};
  \code{\link{load_meta}}; \code{\link{load_nc}}
}

