% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_boxarea.R
\name{load_boxarea}
\alias{load_boxarea}
\title{Extract the box area from the bgm-file!}
\usage{
load_boxarea(file_bgm)
}
\arguments{
\item{prm_biol}{Connection of the ATLANTIS bgm-file file given as complete folder/filename string.
Usually "[...].bgm".}
}
\value{
dataframe with the columns polygon and area. Polygon gives the box id as integer and area
is the box size in m².
}
\description{
This function extracts the box area for each polygon from the atlantis bgm-file.
}
\seealso{
Other load functions: \code{\link{load_acronyms}},
  \code{\link{load_age_acronyms}},
  \code{\link{load_age_groups}},
  \code{\link{load_fish_acronyms}},
  \code{\link{load_groups}},
  \code{\link{load_nonage_acronyms}},
  \code{\link{read_functionalgroups}};
  \code{\link{load_box}}; \code{\link{load_diet_comp}};
  \code{\link{load_meta}}; \code{\link{load_nc}}
}

