% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bps.R
\name{load_bps}
\alias{load_bps}
\title{Extracts the names of the epibenthic biomasspools
from the initial conditions file.}
\usage{
load_bps(dir = getwd(), fgs, file_init)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_init}{A character value giving the file name of the intial conditions
file. The file should be located in your current working directory or the
folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.
Usually the file is named \code{"init[...].nc".}, but it must end in \code{.nc}.}

\item{file_fgs}{A character value specifying the file name of the Atlantis
functional groups file. The file should be located in your current working
directory or the folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.
Usually, the filename is \code{"functionalGroups.csv"}.}
}
\value{
A \code{vector} of biomass pools.
}
\description{
Use \code{file_fgs} \code{data.frame} (this function will call \code{load_fgs})
to get the biomass pool information.
}
\examples{
d <- system.file("extdata", "SETAS_Example", package = "atlantisom")
bps <- load_bps(dir = d, fgs = "Functional_groups.csv", file_init = "Initial_condition.nc")

}
\seealso{
\code{\link{load_fgs}}

Other load functions: 
\code{\link{load_agebioind}()},
\code{\link{load_bioind}()},
\code{\link{load_biolprm}()},
\code{\link{load_boxarea}()},
\code{\link{load_box}()},
\code{\link{load_catch}()},
\code{\link{load_detailed_diet_comp}()},
\code{\link{load_diet_comp}()},
\code{\link{load_fgs}()},
\code{\link{load_fisheries}()},
\code{\link{load_meta}()},
\code{\link{load_nc_annage}()},
\code{\link{load_nc_catchtons}()},
\code{\link{load_nc}()},
\code{\link{load_runprm}()},
\code{\link{load_yoy}()}
}
\author{
Alexander Keth
}
\concept{load functions}
