% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_catch.R
\name{load_catch}
\alias{load_catch}
\title{Load Atlantis \code{[...]catch.txt} file}
\usage{
load_catch(dir, file_catch, fgs, verbose = FALSE)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_catch}{A character value specifying the name of the text file
that contains total annual catch in tonnes output for an Atlantis scenario.
An example entry would be \code{"outputSETAScatch.txt"}.}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. The default value is \code{FALSE}.}
}
\value{
A data frame of total catch weight outputs from the \code{catch.txt} file,
formatted similarly to the outputs of run_truth.
}
\description{
This function loads the Atlantis catch.txt output file.
  Where catch (total annual tonnes) is only available for species that are
  fished in the \code{functionalGroups.csv} file.
}
\examples{
d <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
file <- "outputSETASCatch.txt"
fgs <- load_fgs(dir = d, "functionalGroups.csv")
test <- load_catch(dir = d, fgs = fgs, file_catch = file)
}
\seealso{
Other load functions: 
\code{\link{load_bioind}()},
\code{\link{load_biolprm}()},
\code{\link{load_boxarea}()},
\code{\link{load_box}()},
\code{\link{load_bps}()},
\code{\link{load_diet_comp}()},
\code{\link{load_fgs}()},
\code{\link{load_fisheries}()},
\code{\link{load_meta}()},
\code{\link{load_nc_annage}()},
\code{\link{load_nc}()},
\code{\link{load_runprm}()},
\code{\link{load_yoy}()}
}
\author{
Sarah Gaichas
}
\concept{load functions}
