% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_diet_comp.R
\name{load_diet_comp}
\alias{load_diet_comp}
\title{Load Atlantis Diet Composition}
\usage{
load_diet_comp(dietfile)
}
\arguments{
\item{dietfile}{file name of the diet_check.txt output from Atlantis.}
}
\value{
Returns a data frame of the data to be exported to the AtlantisOM list
 object.
}
\description{
Uses the diet_check.txt output from Atlantis to be used in the Atlantis
Operating model.
}
\seealso{
Other load functions: \code{\link{load_acronyms}},
  \code{\link{load_age_acronyms}},
  \code{\link{load_age_groups}},
  \code{\link{load_fish_acronyms}},
  \code{\link{load_groups}},
  \code{\link{load_nonage_acronyms}},
  \code{\link{read_functionalgroups}};
  \code{\link{load_boxarea}}; \code{\link{load_box}};
  \code{\link{load_meta}}; \code{\link{load_nc}}
}

