% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fgs.R
\name{load_fgs}
\alias{load_fgs}
\title{Load the functional group file}
\usage{
load_fgs(dir = getwd(), file_fgs)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_fgs}{A character value specifying the file name of the Atlantis
functional groups file. The file should be located in your current working
directory or the folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.
Usually, the filename is \code{"functionalGroups.csv"}.}
}
\value{
A \code{data.frame} of functional group information.
}
\description{
Read in the functional group file, which is typically saved as
\code{functionalGroups.csv}.
}
\examples{
d <- system.file("extdata", "SETAS_Example", package = "atlantisom")
fgs <- load_fgs(d, "Functional_groups.csv")
rm(fgs)

}
\seealso{
Other load functions: 
\code{\link{load_bioind}()},
\code{\link{load_biolprm}()},
\code{\link{load_boxarea}()},
\code{\link{load_box}()},
\code{\link{load_bps}()},
\code{\link{load_catch}()},
\code{\link{load_detailed_diet_comp}()},
\code{\link{load_diet_comp}()},
\code{\link{load_fisheries}()},
\code{\link{load_meta}()},
\code{\link{load_nc_annage}()},
\code{\link{load_nc}()},
\code{\link{load_runprm}()},
\code{\link{load_yoy}()}
}
\author{
Alexander Keth
}
\concept{load functions}
