% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fisheries.R
\name{load_fisheries}
\alias{load_fisheries}
\title{Load the fisheries group file}
\usage{
load_fisheries(dir = getwd(), file_fish)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_fish}{A character value specifying the file name of the Atlantis
fisheries file. The file should be located in your current working
directory or the folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.
Usually, the filename includes \code{"fisheries.csv"}.}
}
\value{
A \code{data.frame} of functional group information.
}
\description{
Read in the fisheries group file, which is typically saved as
\code{fisheries.csv}.
}
\examples{
d <- system.file("extdata", "SETAS_Example", package = "atlantisom")
fleets <- load_fisheries(d, "Fisheries.csv")
rm(fleets)

}
\seealso{
Other load functions: 
\code{\link{load_agebioind}()},
\code{\link{load_bioind}()},
\code{\link{load_biolprm}()},
\code{\link{load_boxarea}()},
\code{\link{load_box}()},
\code{\link{load_bps}()},
\code{\link{load_catch}()},
\code{\link{load_detailed_diet_comp}()},
\code{\link{load_diet_comp}()},
\code{\link{load_fgs}()},
\code{\link{load_meta}()},
\code{\link{load_nc_annage}()},
\code{\link{load_nc}()},
\code{\link{load_runprm}()},
\code{\link{load_yoy}()}
}
\author{
Alexander Keth
}
\concept{load functions}
