% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_groups.R
\name{load_groups}
\alias{load_acronyms}
\alias{load_age_acronyms}
\alias{load_age_groups}
\alias{load_fish_acronyms}
\alias{load_groups}
\alias{load_nonage_acronyms}
\alias{read_functionalgroups}
\title{Collection of similar functions which load specific columns from the Atlantis functionalgroups-file}
\usage{
load_groups(file_fgs)

read_functionalgroups(file_fgs)

load_age_groups(file_fgs)

load_acronyms(file_fgs)

load_age_acronyms(file_fgs)

load_nonage_acronyms(file_fgs)

load_fish_acronyms(file_fgs)
}
\arguments{
\item{file_fgs}{Connection of the ATLANTIS functional groups file given as complete folder/filename string.
Usually "functionalGroups.csv".}

\item{nc_init}{Connection of the ATLANTIS init file given as complete folder/filename string. Usually "init[...].nc".}
}
\value{
Character string.
}
\description{
This collection of function loads the Atlantis functional groups file and creates various
character strings of group names or acronym names.
}
\details{
Currently, the following character strings can be created
- load_groups extracts the column "Name"
- load_age_groups extracts the column "Name". Only groups with 10 ageclasses are selected.
- load_acronym extracts the column "Code"
- load_age_acronym extracts the column "Code". Only groups with 10 ageclasses are selected.
- load_nonage_acronym extracts the columne "Code". Only groups with ageclasses different from 10 are selected.
- load_fish_acronyms ectracts the column "Code". Only groups with InvertType equal to "FISH" or "SHARK" are selected.
- load_bps extracts the names of the epibenthic biomasspools from the initial conditions file.
}
\examples{
load_atlantis_output(model_path = file.path("z:", "Atlantis", "ATLANTIS NSmodel base"), filename = "outputNorthSea.nc", select_groups = get_groups(), select_variable = "ResN", biomasspools = c("large_crabs", "small_epifauna", "sessile_epifauna", "epifaunal_macrobenthos"))
}
\seealso{
Other load functions: \code{\link{load_boxarea}};
  \code{\link{load_box}}; \code{\link{load_diet_comp}};
  \code{\link{load_meta}}; \code{\link{load_nc}}
}
\keyword{gen}

