% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nc_annage.R
\name{load_nc_annage}
\alias{load_nc_annage}
\title{Load Atlantis annual age outputfiles (netcdf)}
\usage{
load_nc_annage(
  dir = getwd(),
  file_nc,
  bps,
  fgs,
  biolprm,
  select_groups,
  select_variable = c("Nums", "Weight"),
  check_acronyms = TRUE,
  bboxes = c(0),
  verbose = FALSE
)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_nc}{A character value giving the file name of the netcdf file
you are trying to read in. The file should be located in your current working
directory or the folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.}

\item{bps}{A character vector of biomass pools as read in from
\code{\link{load_bps}}.
The biomass pools are not true biomass pools but all epibenthic
groups from the functional groups file.}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{biolprm}{A list of biological parameters available from the
\code{[...]_Biol.prm} file, as read in by \code{\link{load_biolprm}}.}

\item{select_groups}{Character vector of funtional groups to select.
Names have to match the ones used in the ncdf file, and therefore must
be in the column \code{"Name"} in the \code{fgs}.}

\item{select_variable}{A character value spefifying which variable to return.
loaded. Only one variable of the options available (i.e., \code{c(
"Nums", "Weight")
}) can be loaded at a time.}

\item{check_acronyms}{A logical, specifying if selected groups are active
in the model run. This is used in automated runs.
All groups are passed when plotting is called via batch-file,
(which cannot be changed easily) this will result in errors if some
groups are not active in the model run.
By default this argument is \code{TRUE}.}

\item{bboxes}{A vector of boundary boxes.
The default value is \code{c(0)} because all Atlantis
models must specify box zero as a boundary box.
The boundary boxes can automatically be determined using
\code{get_boundary(load_box())}.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. The default value is \code{FALSE}.}
}
\value{
A \code{data.frame} in long format with the following coumn names:
  Species, timestep, polygon, agecl, and atoutput (i.e., variable).
}
\description{
This function loads Atlantis ANNAGEBIO and ANNAGECATCH outputfiles (netcdf) and
converts them to a \code{data.frame}.
}
\examples{
d <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
fgs <- load_fgs(d, "functionalGroups.csv")
bps <- load_bps(dir = d, fgs = fgs, file_init = "INIT_VMPA_Jan2015.nc")
test <- load_nc(dir = d, file_nc = "outputSETASCATCH.nc",
  fgs = fgs, bps = bps,
  select_variable = "Catch",
  select_groups = "Pisciv_T_Fish",
  check_acronyms = TRUE)
str(test)
rm(test)

}
\seealso{
\code{\link{get_boundary}}, \code{\link{load_box}}

Other load functions: 
\code{\link{load_bioind}()},
\code{\link{load_biolprm}()},
\code{\link{load_boxarea}()},
\code{\link{load_box}()},
\code{\link{load_bps}()},
\code{\link{load_catch}()},
\code{\link{load_diet_comp}()},
\code{\link{load_fgs}()},
\code{\link{load_meta}()},
\code{\link{load_nc}()},
\code{\link{load_runprm}()},
\code{\link{load_yoy}()}
}
\author{
Alexander Keth, modified by Sarah Gaichas
}
\concept{load functions}
\keyword{gen}
