% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nc_physics.R
\name{load_nc_physics}
\alias{load_nc_physics}
\title{Load Atlantis outputfiles (netcdf)}
\usage{
load_nc_physics(
  dir = getwd(),
  file_nc,
  physic_variables,
  aggregate_layers,
  bboxes
)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_nc}{A character value giving the file name of the netcdf file
you are trying to read in. The file should be located in your current working
directory or the folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.}

\item{physic_variables}{A character value spefifying which variables
should be loaded. Only one variable can be loaded at a time.
Currently, only the following variables are allowed:
"salt", "NO3", "NH3", "Temp", "Oxygen", "Si", "Det_Si", "DON",
 "Chl_a", "Denitrifiction", "Nitrification".}

\item{aggregate_layers}{A logical specifying if layers shall be aggregated.}

\item{bboxes}{A vector of boundary boxes.
The default value is \code{c(0)} because all Atlantis
models must specify box zero as a boundary box.
The boundary boxes can automatically be determined using
\code{get_boundary(load_box())}.}
}
\value{
A \code{data.frame} in long format with the following information:
  variable, timestep, polygon, and value (= "atoutput").
}
\description{
This function loads Atlantis outputfiles (netcdf) and converts them to a
\code{data.frame}. The function should be replaced by adding additional
functionality to \code{\link{load_nc}}, but it works as is for right now.
Todo: remove the function and fix \code{\link{load_nc}}
}
\seealso{
\code{\link{get_boundary}}, \code{\link{load_box}}
}
\keyword{gen}
