% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_runprm.R
\name{load_runprm}
\alias{load_runprm}
\title{Load Atlantis \code{_run.prm} file}
\usage{
load_runprm(dir = getwd(), file_runprm)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_runprm}{A character value giving the file name of the biology
\code{.xml} file.
The file should be located in your current working directory or the
folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.
Usually the file is named \code{"[...]_run.xml".}.
This file is output from an Atlantis run in a standard format and is based
on input from the \code{"[...]_run.prm"} file.}
}
\value{
A list of parameters dictating run characteristics from the
\code{_run.xml} file.
}
\description{
This function loads the Atlantis run parameter output file.
}
\examples{
d <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
runprm <- load_runprm(d, "VMPA_setas_run_fishing_F_Trunk.xml")
rm(runprm)

}
\seealso{
Other load functions: \code{\link{load_bioind}},
  \code{\link{load_biolprm}}, \code{\link{load_boxarea}},
  \code{\link{load_box}}, \code{\link{load_bps}},
  \code{\link{load_catch}}, \code{\link{load_diet_comp}},
  \code{\link{load_fgs}}, \code{\link{load_meta}},
  \code{\link{load_nc}}, \code{\link{load_yoy}}
}
\author{
Emma E Hodgson, Kelli Faye Johnson
}
\concept{load functions}
