% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yoy.R
\name{load_yoy}
\alias{load_yoy}
\title{Load Atlantis \code{[...]YOY.txt} file}
\usage{
load_yoy(dir, file_yoy)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_yoy}{A character value specifying the name of the text file
that contains young of the year output for an Atlantis scenario.
An example entry would be \code{"outputSETASYOY.txt"}.}
}
\value{
A data frame of recruitment outputs from the \code{YOY.txt} file.
}
\description{
This function loads the Atlantis young of the year file.
  Where young of the year is only available for species that are turned
  on in the \code{functionalGroups.csv} file.
}
\examples{
d <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
file <- "outputSETASYOY.txt"
test <- load_yoy(dir = d, file_yoy = file)
}
\author{
Emma E Hodgson
}
\seealso{
Other load functions: \code{\link{load_biolprm}},
  \code{\link{load_boxarea}}, \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_diet_comp}},
  \code{\link{load_fgs}}, \code{\link{load_meta}},
  \code{\link{load_nc}}, \code{\link{load_runprm}}
}

