% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_comps.R
\name{om_comps}
\alias{om_comps}
\title{Generate composition data from atlantisom
#'@description A wrapper function to create survey and fishery compositional data for assessment input.
Takes the output of \code{om_species}. Wrapper can generate replicates. Saves output as .rds}
\usage{
om_comps(
  usersurvey = usersurvey_file,
  userfishery = userfishery_file,
  omlist_ss,
  n_reps = n_reps,
  save = TRUE
)
}
\arguments{
\item{usersurvey}{survey config file in format of /config/usersurvey.R}

\item{userfishery}{fishery config file in format of /config/fisherycensus.R}

\item{omlist_ss}{output of \code{om_species}}

\item{n_reps}{number of replicate age, length, and weight-at-age compositions to be generated}

\item{A}{logical value specifying if the function should save the output
to the disk or not. The default is \code{TRUE}.}
}
\value{
Returns list objects containing dataframes of survey biomass index and total catch:
\itemize{
 \item{survObsBiomB, list of replicate dataframes of observed survey biomass (tons)}
 \item{fishObsCatchB, list of replicate dataframes of observed fishery catch (tons)}
 \item{survObsAgeComp, list of replicate dataframes of observed survey age comp (n fish)}
 \item{survObsLenComp, list of replicate dataframes of observed survey length comp (n fish)}
 \item{survObsWtAtAge, list of replicate dataframes of observed survey weight at age (n fish)}
 \item{fishObsAgeComp, list of replicate dataframes of observed fishery age comp (n fish)}
 \item{fishObsLenComp, list of replicate dataframes of observed fishery length comp (n fish)}
 \item{fishObsWtAtAge, list of replicate dataframes of observed fishery weight at age (n fish)}
},
}
\description{
Generate composition data from atlantisom
#'@description A wrapper function to create survey and fishery compositional data for assessment input.
Takes the output of \code{om_species}. Wrapper can generate replicates. Saves output as .rds
}
\examples{
\dontrun{
# assuming CC3om is output of om_init(here("config/CC3config.r"))
# and CC3om_sardine <- om_species(c("Pacific_sardine"), CC3om)

CC3om_sard_comp <- om_comps(usersurvey = here("config/usersurvey.R"),
    userfishery = here("config/fisherycensus.R"),
    omlist_ss = CC3om_sardine,
    n_reps = 1,
    save = TRUE)

}
}
\seealso{
Other wrapper functions: 
\code{\link{om_index}()},
\code{\link{om_init}()},
\code{\link{om_species}()}
}
\author{
Sarah Gaichas
}
\concept{wrapper functions}
