% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_diet.R
\name{om_diet}
\alias{om_diet}
\title{Generate diet composition data from atlantisom
#'@description A wrapper function to create survey diet composition data for model input.
Takes the output of \code{om_species}. Wrapper can generate replicates. Saves output as .rds
Results for more than one survey are generated with multiple survey config files and
saved as separate .rds files.}
\usage{
om_diet(
  config = configfile,
  dietfile = file_diet,
  fgs = fgs,
  usersurvey = usersurvey_file,
  omlist_ss,
  n_reps = n_reps,
  unidprey = 0,
  alphamult = 1e+07,
  save = TRUE
)
}
\arguments{
\item{config}{file specifying directory with atlantis model outputs and identifying key files}

\item{dietfile}{A character value, specifying the file name of the
\code{DetailedDietCheck.txt} output file from Atlantis (or see \code{\link{load_detailed_diet}}).
This file is read with \code{data.table::fread()}, so it can be
compressed with a .gz suffix, or uncompressed as a .txt file.
If \code{is.null(dir)}, then \code{file_diet} can be the full file
path or a file in your current working directory, or the \code{file_diet}
will be appended to \code{dir} using \code{file.path}.}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{usersurvey}{survey config file in format of /config/usersurvey.R}

\item{omlist_ss}{output of \code{om_species}}

\item{n_reps}{number of replicate age, length, and weight-at-age compositions to be generated}

\item{unidprey}{default (0) returns correctly identified prey, while
decimal values up to 1 allocate up to that proportion of prey to an unidentified
category at random.}

\item{alphamult}{default (10000000) returns true diet comps, while a
lower number (~10) returns high variance diet comp}

\item{A}{logical value specifying if the function should save the output
to the disk or not. The default is \code{TRUE}.}
}
\value{
Returns a list object containing dataframes of survey diet proportion:
\itemize{
 \item{survObsDiets, list of replicate dataframes of observed survey diets (proportion)}
},
}
\description{
Generate diet composition data from atlantisom
#'@description A wrapper function to create survey diet composition data for model input.
Takes the output of \code{om_species}. Wrapper can generate replicates. Saves output as .rds
Results for more than one survey are generated with multiple survey config files and
saved as separate .rds files.
}
\examples{
\dontrun{
# assuming CC3om is output of om_init(here("config/CC3config.r"))
# and CC3om_sardine <- om_species(c("Pacific_sardine"), CC3om)

CC3om_sard_comp <- om_comps(usersurvey = here("config/usersurvey.R"),
    userfishery = here("config/fisherycensus.R"),
    omlist_ss = CC3om_sardine,
    n_reps = 1,
    save = TRUE)

}
}
\seealso{
Other wrapper functions: 
\code{\link{om_comps}()},
\code{\link{om_index}()},
\code{\link{om_init}()},
\code{\link{om_species}()}
}
\author{
Sarah Gaichas
}
\concept{wrapper functions}
