% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_index.R
\name{om_index}
\alias{om_index}
\title{Generate index data from atlantisom
#'@description A wrapper function to create survey and fishery index data for assessment input.
Takes the output of \code{om_species}. Wrapper can generate replicates. Saves output as .rds
Results for more than one survey are generated with multiple survey config files and
saved as separate .rds files.}
\usage{
om_index(
  usersurvey = usersurvey_file,
  userfishery = userfishery_file,
  omlist_ss,
  n_reps = n_reps,
  save = TRUE
)
}
\arguments{
\item{usersurvey}{survey config file in format of /config/usersurvey.R}

\item{userfishery}{fishery config file in format of /config/fisherycensus.R}

\item{omlist_ss}{output of \code{om_species}}

\item{n_reps}{number of replicate indices to be generated}

\item{A}{logical value specifying if the function should save the output
to the disk or not. The default is \code{TRUE}.}
}
\value{
Returns list objects containing dataframes of survey biomass index and total catch:
\itemize{
 \item{survObsBiomB, list of replicate dataframes of observed survey biomass (tons)}
 \item{fishObsCatchB, list of replicate dataframes of observed fishery catch (tons)}
},
}
\description{
Generate index data from atlantisom
#'@description A wrapper function to create survey and fishery index data for assessment input.
Takes the output of \code{om_species}. Wrapper can generate replicates. Saves output as .rds
Results for more than one survey are generated with multiple survey config files and
saved as separate .rds files.
}
\examples{
\dontrun{
# assuming CC3om is output of om_init(here("config/CC3config.r"))
# and CC3om_sardine <- om_species(c("Pacific_sardine"), CC3om)

CC3om_sard_ind <- om_index(usersurvey = here("config/usersurvey.R"),
    userfishery = here("config/fisherycensus.R"),
    omlist_ss = CC3om_sardine,
    n_reps = 5,
    save = TRUE)
}

}
\seealso{
Other wrapper functions: 
\code{\link{om_comps}()},
\code{\link{om_diet}()},
\code{\link{om_init}()},
\code{\link{om_species}()}
}
\author{
Sarah Gaichas
}
\concept{wrapper functions}
