% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_init.R
\name{om_init}
\alias{om_init}
\title{Initializes atlantisom}
\usage{
om_init(config = configfile)
}
\arguments{
\item{config}{file specifying directory with atlantis model outputs and identifying key files}
}
\value{
Returns an omlist list object containing dataframes and lists:
\itemize{
 \item{funct.groups, dataframe of species characteristics}
 \item{funct.group.names, vector of species names}
 \item{truetotbio, dataframe output of \code{load_bioind}}
 \item{truecatchbio, dataframe output of \code{load_catch}}
 \item{YOY, dataframe young of year output}
 \item{biol, list of biological parameters output of \code{load_biolprm}}
 \item{runpar, list of run parameters output of \code{load_runprm}}
 \item{boxpars, dataframe o spatial parameters}
 \item{truth, list of dataframes, output of \code{run_truth}}
},
the atlantisom truth and key parameters list object
}
\description{
A wrapper function to initialize an \code{atlantisom} objects from atlantis
output prior to creating assessment input datasets. This makes a large object!
}
\examples{
\dontrun{
library(here)
CC3om <- om_init(here("config/CC3config.r"))

# requires these files in config, for example CC3config.R is:

# d.name <- here::here("atlantisoutput","CC_2063_OA_OFF_22")
# functional.groups.file <- "CalCurrentV3Groups.csv"
# biomass.pools.file <- "DIVCalCurrentV3_Biol.nc"
# biol.prm.file <- "CalCurrentV3_Biol.prm"
# box.file <- "CalCurrentV3_utm.bgm"
# initial.conditions.file <- "DIVCalCurrentV3_Biol.nc"
# run.prm.file <- "CalCurrentV3_run.xml"
# scenario.name <- "CCV3"
# bioind.file <- "outputCCV3BiomIndx.txt"
# catch.file <- "outputCCV3Catch.txt"
# annage <- FALSE
# fisheries.file <- "CalCurrentV3Fisheries.csv"
}

}
\seealso{
Other wrapper functions: 
\code{\link{om_comps}()},
\code{\link{om_diet}()},
\code{\link{om_index}()},
\code{\link{om_species}()}
}
\author{
Sarah Gaichas
}
\concept{wrapper functions}
