% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_species.R
\name{om_species}
\alias{om_species}
\title{Species-specific atlantisom outputs}
\usage{
om_species(
  species = spp,
  omlist,
  save = TRUE,
  diet = FALSE,
  removefullom = TRUE
)
}
\arguments{
\item{species}{The species to sample in the survey and fishery (a vector)}

\item{omlist}{output of \code{om_init}}
}
\value{
Returns an omlist_ss list object containing dataframes and lists:
\itemize{
 \item{species_ss, species name(s)}
 \item{code_ss, species code from funct.groups}
 \item{truetotbio_ss, dataframe output of \code{load_bioind}, species_ss only}
 \item{truecatchbio_ss, dataframe output of \code{load_catch}, species_ss only}
 \item{YOY_ss, dataframe young of year output, species_ss only}
 \item{truenums_ss, numbers at age output of \code{run_truth}, species_ss only}
 \item{truebio_ss, biomass at age output of \code{run_truth}, species_ss only}
 \item{trueresn_ss, reserve nitrogen output of \code{run_truth}, species_ss only}
 \item{truestructn_ss, structural nitrogen output of \code{run_truth}, species_ss only}
 \item{truecatchnum_ss, fishery catch at age output of \code{run_truth}, species_ss only}
 \item{truebioeaten_ss, biomass_eaten output of \code{run_truth}, species_ss only}
 \item{funct.groups_ss, dataframe of species characteristics, species_ss only}
 \item{biol, list of biological parameters passed from input omlist}
 \item{boxpars, dataframe of spatial parameters}
 \item{runpar, list of run parameters passed from input omlist}
},
}
\description{
A wrapper function to reduce a set of \code{atlantisom} objects from atlantis
output to a single species or subset of species for creating assessment input datasets.
Optionally includes diet data (set diet=TRUE in function call, default is FALSE)
}
\examples{
\dontrun{
# assuming CC3om is output of om_init(here("config/CC3config.r"))
CC3om_sardine <- om_species(c("Pacific_sardine"), CC3om)

CC3om_2spp <- om_species(c("Pacific_sardine", "Mesopel_M_Fish"), CC3om)
}

}
\seealso{
Other wrapper functions: 
\code{\link{om_comps}()},
\code{\link{om_index}()},
\code{\link{om_init}()}
}
\author{
Sarah Gaichas
}
\concept{wrapper functions}
