% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_savedsurvs.R
\name{read_savedsurvs}
\alias{read_savedsurvs}
\title{Read saved survey outputs into R object}
\usage{
read_savedsurvs(dir, type)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{type}{A variable passed to \code{read_savedsurvs} to determine the type
of survey output to be read in. Must take one of these values:
  \itemize{
    \item{survB}
    \item{survAge}
    \item{survLen}
    \item{survWtage}
    \item{survAnnAge}
    \item{survAnnWtage}
  }}
}
\description{
A function to read survey index and composition data saved as .rds into
an R object with the same structure as the output of \code{om_species} or \code{om_comps}.
Users must enter the path where the survey output files are stored and the type of survey
output file to be read in.
}
\details{
\code{om_species} and \code{om_comps} allow for multiple surveys. Users
must define a survey.name in each survey config file which is used to name the file
that is output by each survey function, and is then read back in to name list
output for each survey in the saved .rds file.
}
\examples{
\dontrun{
# assuming d.name is the path where atlantisom output is stored

survObsBiom <- read_savedsurvs(d.name, 'survB') # survey biomass index
age_comp_data <- read_savedsurvs(d.name, 'survAge') # survey age class composition
len_comp_data <- read_savedsurvs(d.name, 'survLen') # survey length composition
wtage <- read_savedsurvs(d.name, 'survWtage') # survey weight at age class
annage_comp_data <- read_savedsurvs(d.name, 'survAnnAge') # survey annual age composition
annage_wtage <- read_savedsurvs(d.name, 'survAnnWtage') # survey weight at annual age

}

}
\author{
Sarah Gaichas
}
\concept{read functions}
