% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_compositions.R
\name{reformat_compositions}
\alias{reformat_compositions}
\title{Reformat compositional data for input in to Stock Synthesis}
\usage{
reformat_compositions(comp_data, round.places = 4, comp_type)
}
\arguments{
\item{comp_data}{A \code{data.frame} of numbers, length, or weight at age containing the following columns:
\itemize{
  \item{species}
  \item{agecl}
  \item{polygon}
  \item{layer}
  \item{time}
  \item{atoutput}
  \item{lower.bins}
  \item{upper.bins}
}
The \code{data.frame} is generated from either \code{\link{create_survey}}
or \code{\link{create_fishery_data}}.}

\item{comp_type}{takes 3 possible values: lencomp, agecomp, caalcomp}
}
\value{
flattened comp data set for input into SS
}
\description{
to be added
}
\details{
to be added
}
