% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_atlantis.R
\name{run_atlantis}
\alias{run_atlantis}
\title{Load Atlantis scenario output}
\usage{
run_atlantis(scenario, dir = getwd())
}
\arguments{
\item{scenario}{output name of the Atlantis scenario.}

\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}
}
\value{
Returns a list object.
}
\description{
Reads in data generated from an Atlantis scenario and returns a list
containing the desired information. The list contains the 'truth' as known
from the Atlantis scenario. The truth can later be sampled
from to create a data set with observation error.
Currently, the \code{run_atlantis} depends on the following files
being in your working directory:
\itemize{
 \item{"functionalGroups.csv"}
 \item{"...TOTCATCH.nc"}
 \item{"...DietCheck.txt"}
}
}
\author{
Sean Lucey
}

