% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stocksynthesis.R
\name{run_stocksynthesis}
\alias{run_stocksynthesis}
\title{Function to execute an SS run}
\usage{
run_stocksynthesis(os, species, model_dir, admb_options = " -noest",
  ignore.stdout = TRUE, show_output = FALSE, iteration_pause = 0.05,
  ...)
}
\arguments{
\item{species}{species name}

\item{model_dir}{directory containing the model files}

\item{admb_options}{options for running admb, like "-nohess", "-noest"}

\item{ignore.stdout}{Logical: passed on to
\code{\link[base]{system}}.}

\item{show_output}{Passed to \code{system}. If \code{TRUE} then ADMB
output is not printed on screen. This will be slightly faster. Set to
\code{FALSE} to help with debugging.}

\item{iteration_pause}{how long to pause between model iterations if this function is called in a loop}

\item{...}{Anything else to pass to \code{\link[base]{system}}}

\item{ss_name}{name of the ss model contained within}
}
\description{
This function runs stock synthesis in the specified directory for the model name given. It is copied almost verbatim from \code{ss3sim::run_ss3model()}.
}
