% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ages.R
\name{sample_ages}
\alias{sample_ages}
\title{Add observation error to numbers-at-age data}
\usage{
sample_ages(dat, prop, ageErr = NULL)
}
\arguments{
\item{dat}{A \code{data.frame} of numbers at age containing the following columns:
\itemize{
  \item{species}
  \item{agecl}
  \item{polygon}
  \item{layer}
  \item{time}
  \item{atoutput}
}
The \code{data.frame} is generated from either \code{\link{create_survey}}
or \code{\link{create_fishery_data}}.}

\item{prop}{Percentage of samples for each species: a matrix with nrow=length(species). Columns:
species:  the species name. Matches names in species
prop:     the percentage of age samples for each species (for example, 01 if 1 out of every 10 lengthed fish is sampled for age). Max of 1.}

\item{ageErr}{ageErr is a list with elements containing a matrix for each species
list element name is species name
 matrix rows are true age, columns are assigned age
 species that are missing will be assigned no ageing error (so a value of NULL means no ageing error for all species)}
}
\description{
Create sampled data from numbers-at-age data to add
  observation error to data from and Atlantis scenario.
  todo: add more information here later
}
\details{
The function takes numbers-at-age data from an Atlantis scenario
  where the data was read in from Atlantis output using \code{\link{load_nc}}
  within \code{\link{run_atlantis}}. One does not need to use these functions
  to create \code{dat}, rather you must only ensure that the structure of
  \code{dat} is the same.
  Currently, the function sums across boxes to properly weight the
  numbers-at-age data. Subsequently, the data is sampled using a multinomial
  with the effective sample size based on the number of fish you pass in
  \code{dat} and \code{prop}.
  Additionally, you can apply ageing error to the data, using the \code{ageErr}
  argument.
}
\examples{
		setwd(file.path(system.file( package = "atlantisom"),".."))

	directory <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
	scenario <- "SETAS"
	groups <- load_fgs(dir = directory, "functionalGroups.csv")
	groups <- groups[groups$IsTurnedOn > 0, "Name"]
	results <- run_atlantis(scenario = scenario,
	dir = directory,
	file_fgs = "functionalGroups.csv",
	file_bgm = "VMPA_setas.bgm",
	select_groups = groups,
	file_init = "INIT_VMPA_Jan2015.nc",
	file_biolprm = "VMPA_setas_biol_fishing_Trunk.prm")

	species=c("Pisciv_T_Fish","Pisciv_V_Fish")
    boxes <- 1:3
	effic <- data.frame(species=c("Pisciv_T_Fish","Pisciv_V_Fish"), efficiency=c(0.3,0.1))
	selex <- data.frame(species=c(rep("Pisciv_T_Fish",10),rep("Pisciv_V_Fish",10)),
	                agecl=c(1:10,1:10),
	                selex=c(0,0,0.1,0.5,0.8,1,1,1,1,1,0,0,0.1,0.3,0.5,0.7,0.9,1,1,1))

	tmp <- create_survey(dat=results$nums, time=seq(70,82,3), species=species, boxes=boxes, effic=effic, selex=selex)
	effN <- data.frame(species=species, effN=c(200, 500))
	samp <- sample_fish(tmp, effN=effN)

    prop <- data.frame(species=species, prop=c(0.5,1)) #should be same as input when prop=1, but with ageing error
     sample_ages(samp,prop,ageErr=NULL)
}
\author{
Poseidon
}

