% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fish.R
\name{sample_fish}
\alias{sample_fish}
\title{Sample numbers-at-age to create composition data}
\usage{
sample_fish(dat, effN, sample = TRUE)
}
\arguments{
\item{dat}{A \code{data.frame} of numbers at age containing the following columns:
\itemize{
  \item{species}
  \item{agecl}
  \item{polygon}
  \item{layer}
  \item{time}
  \item{atoutput}
}
The \code{data.frame} is generated from either \code{\link{create_survey}}
or \code{\link{create_fishery_subset}}.}

\item{effN}{Efficiency for each species: a matrix with nrow=length(species). Columns:
species:  the species name. Matches names in species
effN:     the effective N for each species (effective sample size)}

\item{sample}{Logical asking whether to apply multinomial sampling using effN.
Setting to false results in simple aggregation of atoutput to annual age class values.
The default value is \code{TRUE}.}
}
\description{
Create sampled data from numbers-at-age data to create
  composition data from and Atlantis scenario by default.
  todo: add more information here later
}
\details{
The function takes numbers-at-age data from an Atlantis scenario
  where the data was read in from Atlantis output using \code{\link{load_nc}}
  within \code{\link{run_truth}}. One does not need to use these functions
  to create \code{dat}, rather you must only ensure that the structure of
  \code{dat} is the same.
  Currently, the function creates sampled numbers-at-age data to create
  composition data (age, convert to length, age-at-weigth, etc.).
  sums across boxes to properly weight the
  numbers-at-age data. Subsequently, the data is sampled using a multinomial
  with the effective sample size based on the number of fish you pass in
  \code{dat} and \code{prop}. Density data are also needed for biological sampling.
  Therefore, when called with sample=FALSE for density data,
  applies median to aggregate and does not apply multinomial.
  The function starts with numbers-at-age data and therefore the function then must
  sum across boxes to properly weight them. Subsequently, the function then uses
  a multinomial with a specified effective sample size.
  The function could be improved by adding an argument to create spatial
  strata which are conglomerations of subsets of boxes.
}
\examples{
		setwd(file.path(system.file( package = "atlantisom"),".."))

	directory <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
	scenario <- "SETAS"
	groups <- load_fgs(dir = directory, "functionalGroups.csv")
	groups <- groups[groups$IsTurnedOn > 0, "Name"]
	results <- run_truth(scenario = scenario,
	dir = directory,
	file_fgs = "functionalGroups.csv",
	file_bgm = "VMPA_setas.bgm",
	select_groups = groups,
	file_init = "INIT_VMPA_Jan2015.nc",
	file_biolprm = "VMPA_setas_biol_fishing_Trunk.prm")

	species=c("Pisciv_T_Fish","Pisciv_V_Fish")
    boxes <- 1:3
	effic <- data.frame(species=c("Pisciv_T_Fish","Pisciv_V_Fish"), efficiency=c(0.3,0.1))
	selex <- data.frame(species=c(rep("Pisciv_T_Fish",10),rep("Pisciv_V_Fish",10)),
	                agecl=c(1:10,1:10),
	                selex=c(0,0,0.1,0.5,0.8,1,1,1,1,1,0,0,0.1,0.3,0.5,0.7,0.9,1,1,1))

	tmp <- create_survey(dat=results$nums, time=seq(70,82,3), species=species, boxes=boxes, effic=effic, selex=selex)
	effN <- data.frame(species=species, effN=c(200, 500))
	samp <- sample_fish(tmp, effN=effN)

}
\author{
Poseidon
}
