% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fishery_totcatch.R
\name{sample_fishery_totcatch}
\alias{sample_fishery_totcatch}
\title{Sample a fishery catch index from an atlantis scenario}
\usage{
sample_fishery_totcatch(dat, cv)
}
\arguments{
\item{dat}{A \code{data.frame} of numbers at age containing the following columns:
\itemize{
  \item{species}
  \item{agecl}
  \item{polygon}
  \item{layer}
  \item{time}
  \item{atoutput}
}
The \code{data.frame} is generated from either \code{\link{create_survey}}
or \code{\link{create_fishery_data}}.}

\item{cv}{Coefficient of variation for the entire species specific catch
a matrix with columns: species, cv}
}
\value{
The standard dataframe as specified used in \code{dat}.
}
\description{
The function takes total catch data from an Atlantis scenario
  where the data was read in from Atlantis output using \code{\link{load_catch}}
  function. One does not need to use these functions
  to create \code{dat}, rather you must only ensure that the structure of
  \code{dat} is the same.
}
\details{
This function simply creates an observed total catch time series from
  total catch input using by applying a user-specified cv.
  The result is a coastwide catch (tons) estimate in tons from the fishery.
}
\author{
Poseidon
}
