% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_survey_biomass.R
\name{sample_survey_biomass}
\alias{sample_survey_biomass}
\title{Sample a biomass index of abundance from an atlantis scenario}
\usage{
sample_survey_biomass(dat, cv, wtAtAge)
}
\arguments{
\item{dat}{A \code{data.frame} of numbers at age containing the following columns:
\itemize{
  \item{species}
  \item{agecl}
  \item{polygon}
  \item{layer}
  \item{time}
  \item{atoutput}
}
The \code{data.frame} is generated from either \code{\link{create_survey}}
or \code{\link{create_fishery_data}}.}

\item{cv}{Coefficient of variation for the entire species specific biomass
a matrix with columns: species, cv}

\item{wtAtAge}{Weight-at-age by species. a matrix with columns:
species, agecl, time (optional), wtAtAge (kg)}
}
\value{
The standard dataframe as specified used in \code{dat}.
  The function sums over layers and makes \code{$layers} is {NA}.
}
\description{
The function takes numbers-at-age data from an Atlantis scenario
  where the data was read in from Atlantis output using \code{\link{load_nc}}
  within \code{\link{run_truth}}. One does not need to use these functions
  to create \code{dat}, rather you must only ensure that the structure of
  \code{dat} is the same.
}
\details{
This function simply calculates biomass-at-age in t by applying a weight-at-age
  vector (units of kg), sums over polygons,
  and then applies user defined error to the biomass.
  The result is a coastwide biomass estimate in tons from the survey
  Improvements could be to provide polygon specific biomass,
  but the cv will need to be thought about.
}
\examples{
d <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
groups <- load_fgs(dir = d, "functionalGroups.csv")
run_truth(scenario = "SETAS",
  dir = d,
  file_fgs = "functionalGroups.csv",
  file_bgm = "VMPA_setas.bgm",
  select_groups = groups[groups$IsTurnedOn > 0, "Name"],
  file_init = "INIT_VMPA_Jan2015.nc",
  file_biolprm = "VMPA_setas_biol_fishing_Trunk.prm",
  file_runprm = "VMPA_setas_run_fishing_F_Trunk.xml")

species=c("Pisciv_T_Fish","Pisciv_S_Fish")
 boxes <- 1:3
effic <- data.frame(species=c("Pisciv_T_Fish","Pisciv_S_Fish"), efficiency=c(0.3,0.1))
selex <- data.frame(species=c(rep("Pisciv_T_Fish",10),rep("Pisciv_S_Fish",10)),
               agecl=c(1:10,1:10),
                selex=c(0,0,0.1,0.5,0.8,1,1,1,1,1,0,0,0.1,0.3,0.5,0.7,0.9,1,1,#'1))

tmp <- create_survey(dat=results$nums, time=seq(70,82,3), species=species, boxes=boxes, effic=effic, selex=sel#'ex)

wtAtAge <- data.frame(species=c(rep("Pisciv_T_Fish",10),rep("Pisciv_S_Fish",10)),
                agecl=c(1:10,1:10),
                wtAtAge=c(0.1,0.5,0.9,1.5,1.8,2.0,2.1,2.2,2.3,2.35,0.05,0.2,0.3,0.35,0.45,0.5,0.55,0.6,0.63,0.65))
cv <- data.frame(species=species, cv=c(0.2,0.3))

survObsBiom <- sample_survey_biomass(dat=tmp,cv=cv,wtAtAge)
}
\author{
Poseidon
}
