% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_write_biol.R
\name{SS_write_biol}
\alias{SS_write_biol}
\title{Writes life history parameters}
\usage{
SS_write_biol(
  ctl_obj,
  biolprm_object,
  species_code,
  M_est = NULL,
  wtsage,
  lensage,
  lenwt_a = NULL,
  lenwt_b = NULL
)
}
\arguments{
\item{ctl_obj}{list that is returned from \code{r4ss::read_ctl()} that contains the control file}

\item{biolprm_object}{list with a number of atlantis biological parameters}

\item{species_code}{the three-letter species code of the fish you are creating a model for}

\item{M_est}{scalar value for M of the species}

\item{wtsage}{two dimensional vector with column names agecl and meanwt giving mean body weight at age in grams}

\item{lenage}{three dimensional vector with column names agecl, meanln, and cvln giving mean body length and cv of body length at age}
}
\value{
an updated ctl_obj that has replaced parameters with the atlantis-created ones
}
\description{
Function to pull relevant life history parameters from atlantis to write to a Stock Synthesis control and/or data file.
}
