% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_write_comps.R
\name{SS_write_comps}
\alias{SS_write_comps}
\title{Function to write age and length composition data from \code{atlantisom} to Stock Synthesis 3.30.}
\usage{
SS_write_comps(ss_data_list, comp_matrix, data_rows, sampling_month,
  data_type, fleet_number, bins, caal_bool)
}
\arguments{
\item{ss_data_list}{the list object containing SS data}

\item{comp_matrix}{the matrix of composition data (age or length)}

\item{data_rows}{a list, where each entry is a vector corresponding to the year column of the composition data. For length composition data, this is the 1:nyears repped each row for the number of length bins. For age composition data, this is nyears}

\item{sampling_month}{a list where each list item is a vector of the sampling month for each composition data type}

\item{data_type}{a vector of strings, each entry needs to be either "agecomp" or "lengthcomp"}

\item{fleet_number}{a vector of numbers, each entry corresponds to which fleet the comp data are from}

\item{bins}{a list item, each item is a vector specifying either the age or length bins}

\item{caal_bool}{a vector of length \code{ss_data_list} with boolean values for whether the data is conditional age-at-length}
}
\description{
Function to write age and length composition data from \code{atlantisom} to Stock Synthesis 3.30.
}
