% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_write_CPUE.R
\name{SS_write_CPUE}
\alias{SS_write_CPUE}
\title{Function to write the CPUE from an atlantis OM into a
Stock Synthesis 3.3 estimation model}
\usage{
SS_write_CPUE(ss_data_list, cpue_data, CVs, data_years, sampling_month,
  units)
}
\arguments{
\item{ss_data_list}{the list of the SS .dat file that #'comes from \code{SS_readdat}}

\item{cpue_data}{a list, where each list item is a #'vector of the observations for each index}

\item{CVs}{a vector of numbers representing the CV around the index}

\item{data_years}{a vector, where each item is the length of the index}

\item{sampling_month}{a list, where each item is a vector representig the month of sampling for each index}

\item{units}{must be one of "numbers" or "biomass"}
}
\description{
Function to write the CPUE from an atlantis OM into a
Stock Synthesis 3.3 estimation model
}
