% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_meta.R
\name{write_meta}
\alias{write_meta}
\title{Write metadata to the disk.}
\usage{
write_meta(meta, dir = getwd(), file_out = "atlantisom_log.txt",
  verbose = FALSE)
}
\arguments{
\item{meta}{A \code{list} of metadata as read in by \code{\link{load_meta}}.}

\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_out}{A character value specifying a file name to write
output to. If the function takes the argument \code{dir}, then
the file will be written to \code{file.path(dir, file_out)}. If the
argument \code{dir} is not included in the function then the file will
be written to the current directory or one can specify a full path name.
The default will depend on the function and if \code{is.null(file_out)}
then no file will be written.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. The default value is \code{FALSE}.}
}
\value{
Write a file to the disk of metadata pertaining to a single Atlantis scenario.
}
\description{
Write metadata to the disk.
}
\author{
Kelli Faye Johnson
}
\seealso{
\code{\link{load_meta}}
}

