% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_SS_ctl.R
\name{write_SS_ctl}
\alias{write_SS_ctl}
\title{write SS control}
\usage{
write_SS_ctl(wt_len_obj, biolprm_obj, ...)
}
\arguments{
\item{wt_len_obj}{??}

\item{biolprm_obj}{??}
}
\description{
Function to write SS control file changes in line with Atlantis operating model.
The function modifies the key biological and selectivity parameter initial values and bounds.
These parameters include: LN(R0), h, M, L1, L2, k, CV_young, CV_old
TO DO: add in logic to modify which parameters are fixed and which are estimated
}
