% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_SS_dat.R
\name{write_SS_dat}
\alias{write_SS_dat}
\title{Read SS dat}
\usage{
write_SS_dat(
  model_dir,
  datfile_name,
  survey_ts,
  fishery_ts,
  survey_config,
  fishery_config,
  fleets = c(2, 1),
  survey_lencomp,
  fishery_lencomp
)
}
\arguments{
\item{model_dir}{the directory (relative to inst extdata) your SS files are in}

\item{datfile_name}{the name of your stock synthesis .dat file}

\item{survey_ts}{a data frame with columns \code{atoutput} and \code{time} where atoutput is survey CPUE in biomass (tons)}

\item{fishery_ts}{a data frame with columns \code{atoutput} and \code{time} where atoutput is fished catch in biomass (tons)}

\item{survey_config}{the name of your survey config fils}

\item{fishery_config}{the name of your fishery config file}

\item{fleets}{a vector corresponding to the index of your fishing fleet followed by the index of your survey fleet}

\item{survey_lencomp}{output from the atlantisom \code{calc_age2length()} function for the survey}

\item{fishery_lencomp}{output from the atlantisom \code{calc_age2length()} function for the fishery}
}
\value{
stocksynthesis.data the dat file list object
}
\description{
A wrapper function to read in an SS .dat file, modify it to add a catch vector, survey CPUE, CAAL and length comp data
from a survey, and age and length comp data from a fishery using the \code{atlantisom} functions to take data from atlantis.
}
