% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregate and select data from Atlantis output}
\usage{
aggregateData(dat, time, species, boxes, keepColumns = c("species", "agecl",
  "polygon", "time"))
}
\arguments{
\item{dat}{The dataframe of numbers-at-age from reading in the Atlantis fileSnapshot}

\item{time}{The timing of the survey (can be a vector if survey spans multiple time periods)}

\item{species}{The species to sample in the survey (a vector)}

\item{boxes}{A vector of box numbers}
}
\value{
The function returns a subsetted matrix with the same columns
  as the input data, i.e.,:
  \itemize{
    \item{species}
    \item{agecl}
    \item{polygon}
    \item{layer}
    \item{time}
    \item{atoutput}
  }
}
\description{
subsets data based on defined columns, and sums over 
layers, but enter \code{NA} for layers to
  keep the column names the same.
  Currently, the function works for a vector of defined species and polygons.
}
\author{
Poseidon
}

