% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_age2length.R
\name{calc_age2length}
\alias{calc_age2length}
\title{Code to get size at age and plot}
\usage{
calc_age2length(natage, structn, reserven, ncfile, biolfile, groupfile,
  CVlenage = 0.01)
}
\arguments{
\item{natage}{dataframe of numbers at age by species, age class, box, depth,
and time}

\item{structn}{dataframe of structural n by species, age class, box, depth,
and time}

\item{reserven}{dataframe of reserve n by species, age class, box, depth,
and time}

\item{ncfile}{full path and filename of biology .prm file}

\item{biolfile}{full path and filename of biology .prm file}

\item{groupfile}{full path and filename of functional groups .csv file
The above should  be changed to only pass the pieces from these that
we need.}

\item{CVlenage}{The variability in length at age (currently same for all species)}
}
\description{
Code to get size at age and plot
}
\author{
Gavin Fay
calculate length composition from Atlantis output (st)age data
Uses numbers at age, plus weight at age, and weight-length relationships to
generate size comps
}

