% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age2length.R
\name{calc_age2length}
\alias{calc_age2length}
\title{Code to get size at age and plot}
\usage{
calc_age2length(natage, structn, reserven, nums, biolprm, ncfile, fgs,
  CVlenage = 0.01)
}
\arguments{
\item{natage}{dataframe of numbers at age by species, age class, box, depth,
and time}

\item{structn}{dataframe of structural n by species, age class, box, depth,
and time}

\item{reserven}{dataframe of reserve n by species, age class, box, depth,
and time}

\item{nums}{A \code{data.frame} containing the numbers per functional-group
("species") per timesetp, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "Nums"}.}

\item{biolprm}{A list of biological parameters available from the
\code{[...]_Biol.prm} file, as read in by \code{\link{load_biolprm}}.}

\item{ncfile}{full path and filename of biology .prm file}

\item{fgs}{A data frame created by \code{\link{read_functionalgroups}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{CVlenage}{The variability in length at age (currently same for all species)}

\item{resn}{A \code{data.frame} containing the reserve nitrogen per functional-group
("species") per timestep, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "ResN"}.}

\item{structn}{A \code{data.frame} containing the structural nitrogen per functional-group
("species") per timestep, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "StructN"}.}
}
\description{
Code to get size at age and plot
}
\details{
The above should  be changed to only pass the pieces from these that
we need.
}
\author{
Gavin Fay
calculate length composition from Atlantis output (st)age data
Uses numbers at age, plus weight at age, and weight-length relationships to
generate size comps
}

