% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age2length.R
\name{calc_age2length}
\alias{calc_age2length}
\title{Calculate length from ages}
\usage{
calc_age2length(structn, resn, nums, biolprm, fgs, CVlenage = 0.1,
  remove.zeroes = TRUE)
}
\arguments{
\item{structn}{A \code{data.frame} containing the structural nitrogen per functional-group
("species") per timestep, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "StructN"}.}

\item{resn}{A \code{data.frame} containing the reserve nitrogen per functional-group
("species") per timestep, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "ResN"}.}

\item{nums}{A \code{data.frame} containing the numbers per functional-group
("species") per timesetp, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "Nums"}.}

\item{biolprm}{A list of biological parameters available from the
\code{[...]_Biol.prm} file, as read in by \code{\link{load_biolprm}}.}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{CVlenage}{The variability in length at age (currently same for all species)}

\item{remove.zeroes}{Logical asking whether to only return numbers at length for
combinations of species, age, box, depth, etc that have numbers>0.
The default value is \code{TRUE}.}
}
\value{
A \code{list} containing three \code{data.frame}s, mulen (mean length at age),
muweight (mean weight at age), and natlength (numbers at length). natlength is in the
same format as other dataframes in the atlantisom package except has two additional
columns to include the length bin information.
}
\description{
calculate length composition from Atlantis output (st)age data
Uses numbers at age, plus weight at age, and weight-length relationships to
generate size comps
}
\examples{

directory <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
load(file.path(directory, "outputSETASrun_atlantis.RData"))

lenout <- calc_age2length(
  structn=result$structn, resn=result$resn, nums=result$nums,
  biolprm = result$biolprm, fgs = result$fgs,
  CVlenage = 0.1, remove.zeroes=TRUE)

}
\author{
Gavin Fay
}

