% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biomass_age.R
\name{calc_biomass_age}
\alias{calc_biomass_age}
\title{Calculate biomass in t per age-structured-groups per polygon, time and ageclass.}
\usage{
calc_biomass_age(nums, resn, structn, biolprm)
}
\arguments{
\item{nums}{A \code{data.frame} containing the numbers per functional-group
("species") per timesetp, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "Nums"}.}

\item{resn}{A \code{data.frame} containing the reserve nitrogen per functional-group
("species") per timestep, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "ResN"}.}

\item{structn}{A \code{data.frame} containing the structural nitrogen per functional-group
("species") per timestep, ageclass, layer and polygon.
The \code{data.frame} has to origininate from \code{\link{load_nc}} using
\code{select_variable = "StructN"}.}

\item{biolprm}{A list of biological parameters available from the
\code{[...]_Biol.prm} file, as read in by \code{\link{load_biolprm}}.}
}
\value{
A \code{data.frame} in tidy format with the following columns:
  species, agecl, time, polygon and atoutput.
  Atoutput is the total biomass in t.
  At this point biomass was aggregated (sum) per layer!
}
\description{
Calculate biomass in t per age-structured-groups per polygon, time and ageclass.
}
\details{
This functions converts the ATLANTIS output to a \code{data.frame}
  which can be processed in R.
}
\examples{
# Set up the example with input files
d <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
fgs <- load_fgs(d, "functionalGroups.csv")
bps <- load_bps(dir = d, fgs = fgs,
  file_init = paste0(tail(strsplit(d, "/")[[1]], 1), ".nc"))
runprm <- load_runprm(d, "VMPA_setas_run_fishing_F_Trunk.xml")
biolprm <- load_biolprm(dir = d,
  file_biolprm = "VMPA_setas_biol_fishing_Trunk.prm")
boxes <- get_boundary(load_box(dir = d, file_bgm = "VMPA_setas.bgm"))

# Get the catch values
catch <- load_nc(dir = d, file_nc = "outputSETASCATCH.nc",
  bps = bps, fgs = fgs, select_groups = fgs[fgs$IsTurnedOn > 0, "Name"],
  select_variable = "Catch", check_acronyms = TRUE, bboxes = boxes)
structn <- load_nc(dir = d, file_nc = "outputSETAS.nc",
  bps = bps, fgs = fgs, select_groups = fgs[fgs$IsTurnedOn > 0, "Name"],
  select_variable = "StructN", check_acronyms = TRUE, bboxes = boxes)
resn <- load_nc(dir = d, file_nc = "outputSETAS.nc",
  bps = bps, fgs = fgs, select_groups = fgs[fgs$IsTurnedOn > 0, "Name"],
  select_variable = "ResN", check_acronyms = TRUE, bboxes = boxes)

biomassatage <- calc_biomass_age(nums = catch,
  resn = resn, structn = structn, biolprm = biolprm)

}
\seealso{
Other calc functions: 
\code{\link{calc_Z}()},
\code{\link{calc_avgwtstage2age}()},
\code{\link{calc_pred_diet}()},
\code{\link{calc_stage2age}()},
\code{\link{calc_timestep2time}()}
}
\author{
Alexander Keth
}
\concept{calc functions}
\keyword{gen}
