% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fishery_data.R
\name{create_fishery_data}
\alias{create_fishery_data}
\title{create_fishery_data R function for an Atlantis scenario}
\usage{
create_fishery_data(dat, time, species, boxes)
}
\arguments{
\item{dat}{The dataframe of numbers-at-age from reading in the Atlantis fileSnapshot}

\item{time}{The timing of the survey (can be a vector if survey spans multiple time periods)}

\item{species}{The species to sample in the survey (a vector)}

\item{boxes}{A vector of box numbers}

\item{effic}{Efficiency for each species: a matrix with nrow=length(species). Columns:
 species:    the species name. Matches names in species
efficiency:}

\item{selex}{Selectivity at age. A dataframe defining selectivity at age for each species. Columns are:
species: the species name. Matches names in species
agecl:   the age class that selectivity represents
selex:   the proportion selected relative to fully selected age classes (between 0 and 1)}
}
\description{
create_fishery_data R function for an Atlantis scenario
}
\details{
Returns a matrix in same format as the input dat matrix

columns: species, agecl, polygon, layer, time, atoutput

--will sum over layers, but enter NA as layer to indicate all layers

This function is for a vector of defined species

Returns only boxes where fishery was sampled
}
\author{
Poseidon
}

