% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey.R
\name{create_survey}
\alias{create_survey}
\title{Create survey data from an Altantis scenario}
\usage{
create_survey(dat, time, species, boxes, effic, selex)
}
\arguments{
\item{dat}{A \code{data.frame} of numbers at age containing the following columns:
\itemize{
  \item{species}
  \item{agecl}
  \item{polygon}
  \item{layer}
  \item{time}
  \item{atoutput}
}
The \code{data.frame} is generated from either \code{\link{create_survey}}
or \code{\link{create_fishery_data}}.}

\item{time}{The timing of the survey (a vector indicating specific time steps, which are typically associated with years)
i.e., seq(365,10*3650,365) would be an annual survey for 10 years}

\item{species}{The species to sample in the survey (a vector)}

\item{boxes}{A vector of polygon numbers that are sampled by survey}

\item{effic}{Efficiency for each species: a matrix with nrow=length(species). Columns:
 species:    the species name. Matches names in species
efficiency:}

\item{selex}{Selectivity at age. A dataframe defining selectivity at age for each species. Columns are:
species: the species name. Matches names in species
agecl:   the age class that selectivity represents
selex:   the proportion selected relative to fully selected age classes (between 0 and 1)}

\item{spex}{The specifications of Atlantis model (box-specific area, habitat, etc.)
for now, from load_boxarea a dataframe with polygon and area column names}

\item{boxes}{A matrix with two columns:
1) polygon:  box ID's that are sampled
2) survArea: area sampled in that box}
}
\value{
A \code{matrix} in the same format as the \code{dat} summed
  over layers, with \code{NA} in the layer column.
}
\description{
Creates survey data from an Atlantis scenario where the survey
  has its own selectivity.
  todo: add more information here.
}
\details{
The function works for a vector of defined species
  and only returns information in boxes that are sampled.
}
\author{
Poseidon
}

