% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_box.R
\name{load_box}
\alias{load_box}
\title{Load the box specification file for an Atlantis scenario}
\usage{
load_box(dir = getwd(), file_bgm)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_bgm}{A character value specifying the file name of a \code{.bgm}
file, which contains information on box locations for an Atlantis scenario.
The file should be located in your current working directory or the
folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.}
}
\value{
A list of information regarding boxes for an Atlantis scenario.
}
\description{
Read in the \code{.bgm} file that specifies the the box coordinates
for an Atlantis scenario.
}
\examples{
d <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
boxes <- load_box(dir = d, file_bgm = "VMPA_setas.bgm")
rm("boxes")

}
\seealso{
Other load functions: \code{\link{load_biolprm}},
  \code{\link{load_boxarea}}, \code{\link{load_bps}},
  \code{\link{load_catch}}, \code{\link{load_diet_comp}},
  \code{\link{load_fgs}}, \code{\link{load_meta}},
  \code{\link{load_nc}}, \code{\link{load_runprm}},
  \code{\link{load_yoy}}
}
\author{
Kelli Faye Johnson
}
\concept{load functions}
