% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_boxarea.R
\name{load_boxarea}
\alias{load_boxarea}
\title{Extract the box area from the bgm-file}
\usage{
load_boxarea(dir = getwd(), file_bgm)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_bgm}{A character value specifying the file name of a \code{.bgm}
file, which contains information on box locations for an Atlantis scenario.
The file should be located in your current working directory or the
folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.}
}
\value{
dataframe with the columns polygon and area.
  Polygon gives the box id as integer and area is the box size in m^2.
}
\description{
This function extracts the box area for each polygon from the atlantis bgm-file.
}
\examples{
d <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
boxarea <- load_boxarea(dir = d, file_bgm = "VMPA_setas.bgm")
rm("boxarea")

}
\seealso{
Other load functions: \code{\link{load_biolprm}},
  \code{\link{load_box}}, \code{\link{load_bps}},
  \code{\link{load_catch}}, \code{\link{load_diet_comp}},
  \code{\link{load_fgs}}, \code{\link{load_meta}},
  \code{\link{load_nc}}, \code{\link{load_runprm}},
  \code{\link{load_yoy}}
}
\concept{load functions}
