% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bps.R
\name{load_bps}
\alias{load_bps}
\title{Extracts the names of the epibenthic biomasspools from the initial conditions file.}
\usage{
load_bps(dir = getwd(), fgs, file_init)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{fgs}{A data frame created by \code{\link{read_functionalgroups}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}

\item{file_init}{A character value giving the file name of the intial conditions
file. The file should be located in your current working directory or the
folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.
Usually the file is named \code{"init[...].nc".}, but it must end in \code{.nc}.}
}
\value{
A \code{vector} of biomass pools.
}
\description{
Use \code{\link{fgs}} \code{data.frame} as read in by \code{\link{load_fgs}}
to get the biomass pool information.
}
\examples{
testscenario <- "INIT_VMPA_Jan2015"
d <- system.file("extdata", testscenario, package = "atlantisom")
file <- "functionalGroups.csv"
fgs <- load_fgs(d, file)
bps <- load_bps(dir = d, fgs = fgs, file_init = "INIT_VMPA_Jan2015.nc")

}
\author{
Alexander Keth
}
\seealso{
\code{\link{load_fgs}}

Other load functions: \code{\link{load_biolprm}},
  \code{\link{load_boxarea}}, \code{\link{load_box}},
  \code{\link{load_diet_comp}}, \code{\link{load_fgs}},
  \code{\link{load_meta}}, \code{\link{load_nc}}
}

