% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_diet_comp.R
\name{load_diet_comp}
\alias{load_diet_comp}
\title{Load Atlantis Diet Composition}
\usage{
load_diet_comp(dietfile, path = ".")
}
\arguments{
\item{dietfile}{file name of the diet_check.txt output from Atlantis.}

\item{path}{location of the dietfile.}
}
\value{
Returns a data frame of the data to be exported to the AtlantisOM list
 object.
}
\description{
Uses the diet_check.txt output from Atlantis to be used in the Atlantis
Operating model.
}

