% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_diet_comp.R
\name{load_diet_comp}
\alias{load_diet_comp}
\title{Load Atlantis Diet Composition}
\usage{
load_diet_comp(dir = getwd(), dietfile, fgs)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{dietfile}{file name of the diet_check.txt output from Atlantis.}

\item{fgs}{A data frame created by \code{\link{read_functionalgroups}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}
}
\value{
Returns a data frame of the data to be exported to the AtlantisOM list
 object.
}
\description{
Uses the diet_check.txt output from Atlantis to be used in the Atlantis
Operating model.
}
\seealso{
Other load functions: \code{\link{load_biolprm}},
  \code{\link{load_boxarea}}, \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_fgs}},
  \code{\link{load_meta}}, \code{\link{load_nc}}
}

