% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_meta.R
\name{load_meta}
\alias{load_meta}
\title{Open the necessary files to document the atlantis model used
for generating data from the scenario}
\usage{
load_meta(dir = getwd(), scenario, verbose = FALSE)
}
\arguments{
\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{scenario}{A character value, specifying the Atlantis scenario name.
More specifically the name which Atlantis uses to name all of the
output files. The value will be used to search for files specified in
\code{dir}, if \code{dir} is also a function argument.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. The default value is \code{FALSE}.}
}
\value{
A \code{list} of metadata pertaining to a single Atlantis scenario.
}
\description{
Open the necessary files to document the atlantis model used
for generating data from the scenario
}
\author{
Kelli Faye Johnson
}
\seealso{
\code{\link{write_meta}}

Other load functions: \code{\link{load_biolprm}},
  \code{\link{load_boxarea}}, \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_diet_comp}},
  \code{\link{load_fgs}}, \code{\link{load_nc}},
  \code{\link{load_runprm}}
}

