% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_compositions.R
\name{reformat_compositions}
\alias{reformat_compositions}
\title{Reformat compositional data for input in to Stock Synthesis}
\usage{
reformat_compositions(comp_data, round.places = 4, comp_type)
}
\arguments{
\item{comp_data}{the data frame of composition data}

\item{round.places}{the number of decimal places you would like your compositions rounded to; default is 4}

\item{comp_type}{takes 3 possible values: lencomp, agecomp, caalcomp}
}
\value{
flattened comp data set for input into SS
}
\description{
to be added
}
\details{
to be added
}
