% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_truth.R
\name{run_truth}
\alias{run_truth}
\title{Load Atlantis scenario output}
\usage{
run_truth(scenario, dir = getwd(), file_fgs, file_bgm, select_groups,
  file_init, file_biolprm)
}
\arguments{
\item{scenario}{A character value, specifying the Atlantis scenario name.
More specifically the name which Atlantis uses to name all of the
output files. The value will be used to search for files specified in
\code{dir}, if \code{dir} is also a function argument.}

\item{dir}{The directory of the atlantis model output, where the
default is \code{getwd()}.}

\item{file_fgs}{A character value specifying the file name of the Atlantis
functional groups file. The file should be located in your current working
directory or the folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.
Usually, the filename is \code{"functionalGroups.csv"}.}

\item{file_bgm}{A character value specifying the file name of a \code{.bgm}
file, which contains information on box locations for an Atlantis scenario.
The file should be located in your current working directory or the
folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.}

\item{select_groups}{Character vector of funtional groups to select.
Names have to match the ones used in the ncdf file, and therefore must
be in the column \code{"Name"} in the \code{fgs}.}

\item{file_init}{A character value giving the file name of the intial conditions
file. The file should be located in your current working directory or the
folder you specify in \code{dir}.
The argument can also be specified as the full path name, just as long as
argument \code{dir} is specified as \code{NULL}.
Usually the file is named \code{"init[...].nc".}, but it must end in \code{.nc}.}
}
\value{
Returns a list object.
}
\description{
Reads in data generated from an Atlantis scenario and returns a list
containing the desired information. The list contains the 'truth' as known
from the Atlantis scenario. The truth can later be sampled
from to create a data set with observation error.
Currently, the \code{run_truth} depends on the following files
being in your working directory:
\itemize{
 \item{"functionalGroups.csv"}
 \item{"[...]TOTCATCH.nc"}
 \item{"[...]DietCheck.txt"}
},
where [...] specifies the entry used for the \code{scenario} argument.
}
\examples{
#load_all()
directory <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
scenario <- "SETAS"
groups <- load_fgs(dir = directory, "functionalGroups.csv")
groups <- groups[groups$IsTurnedOn > 0, "Name"]
run_truth(scenario = scenario,
  dir = directory,
  file_fgs = "functionalGroups.csv",
  file_bgm = "VMPA_setas.bgm",
  select_groups = groups,
  file_init = "INIT_VMPA_Jan2015.nc",
  file_biolprm = "VMPA_setas_biol_fishing_Trunk.prm")

}
\author{
Sean Lucey
}

