% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_diet.R
\name{sample_diet}
\alias{sample_diet}
\title{Sample total consumption to create diet composition data}
\usage{
sample_diet(dat, fgs)
}
\arguments{
\item{dat}{A \code{data.frame} containing sampled predator species
identifiers  in the first
column and prey species consumption proportions in the remaining columns.}

\item{fgs}{A data frame created by \code{\link{load_fgs}} that
reads in the csv file containing functional group names, usually
\code{"functionalGroups.csv"}.}
}
\description{
Create sampled diet composition data from the total consumption
   in an Atlantis scenario. Observation error and bias are added.
}
\details{
The function takes total consumption data from an Atlantis scenario
  where the data was read in from Atlantis output using \code{???}. One does
  not need to use these functions to create \code{dat}, rather you must only
  ensure that the structure of \code{dat} is the same.
  Currently, the function creates sampled diet composition by removing
  non-sampled and non-enumerated species groups from the total
  consumption table from Atlantis.
  Bias is added by setting infrequently consumed (<0.25) prey groups to
  zero at random.
  Error is incorporated into proportional composition entries by
  adding uniform error to true values to half of the observations.
  The function adjusts the remaining table so each row sums to one
  before returning the "observed" mean diet composition summary table.
  The function needs to be generalized to any Atlantis system by
  selecting common species group identifiers to remove from the table.
  Also, more realistic observation error and bias distributions
  could be applied to obtain realistic diet composition data.
}
\examples{
		setwd(file.path(system.file( package = "atlantisom"),".."))

	d <- system.file("extdata", "INIT_VMPA_Jan2015", package = "atlantisom")
	groups <- load_fgs(dir = directory, "functionalGroups.csv")
	groups <- groups[groups$IsTurnedOn > 0, "Name"]
	results <- run_truth(scenario = "SETAS",
	dir = d,
	file_fgs = "functionalGroups.csv",
	file_bgm = "VMPA_setas.bgm",
	select_groups = groups,
	file_init = "INIT_VMPA_Jan2015.nc",
	file_biolprm = "VMPA_setas_biol_fishing_Trunk.prm")
	# rows should each sum to one:
	rowSums(dat[,2:NCOL(dat)])
	dim(dat)

	obsDietComp <- sample_diet(dat)
	dim(obsDietComp)

}
\author{
Robert Wildermuth
}

