% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fishery_totcatch.R
\name{sample_fishery_totcatch}
\alias{sample_fishery_totcatch}
\title{Sample a fishery catch index from an atlantis scenario}
\usage{
sample_fishery_totcatch(dat, cv)
}
\arguments{
\item{dat}{A \code{data.frame} of numbers at age containing the following columns:
\itemize{
  \item{species}
  \item{agecl}
  \item{polygon}
  \item{layer}
  \item{time}
  \item{atoutput}
}
The \code{data.frame} is generated from either \code{\link{create_survey}}
or \code{\link{create_fishery_subset}}.}

\item{cv}{Coefficient of variation for the entire species specific catch
a matrix with columns: species, cv}
}
\value{
The standard dataframe as specified used in \code{dat}.
}
\description{
The function takes total catch data from an Atlantis scenario
  where the data was read in from Atlantis output using \code{\link{load_catch}}
  function or from the \code{\link{load_nc_catchtons}} function.
  One does not need to use these functions
  to create \code{dat}, rather you must only ensure that the structure of
  \code{dat} is the same.
}
\details{
If a total catch time series is input from \code{\link{load_catch}},
  this function simply creates an observed total catch time series from
  total catch input using by applying a user-specified cv.
  If catch by polygon and timestep is input from \code{\link{load_nc_catchtons}}
  followed by \code{\link{create_fishery_subset}}, then catches are aggregated
  over polygons and the subannual timesteps are retained. Fleet specific catch
  can be created by selecting a subset of fleets as input to
  \code{\link{create_fishery_subset}}. Either way, the result is a coastwide
  catch (tons) estimate in tons from the fishery.
}
\examples{
d <- system.file("extdata", "SETAS_Example", package = "atlantisom")
groups <- c("Pisciv_T_Fish","Pisciv_S_Fish")
file <- "outputsCatch.txt"
fgs <- load_fgs(dir = d, "Functional_groups.csv")
truetotcatch <- load_catch(dir = d, fgs = fgs, file_catch = file)
groupcatch <- truetotcatch[truetotcatch$species \%in\% groups, ]
cv <- data.frame(species=groups, cv=c(0.2,0.3))
fishObsCatch <- sample_fishery_totcatch(dat=groupcatch,cv=cv)
}
\author{
Poseidon
}
