% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_write_EWAA.R
\name{SS_write_EWAA}
\alias{SS_write_EWAA}
\title{Function to write an empirical weight at age SS file (wt_at_age.ss) from Atlantis output}
\usage{
SS_write_EWAA(age_classes, years, fleets, seas, catch_mean_wt_matrix,
  surv_mean_wt_matrix, fleet_month, mat_vector)
}
\arguments{
\item{age_classes}{the vector of age classes in the model}

\item{catch_mean_wt_matrix}{mu_weight object in grams}

\item{surv_mean_wt_matrix}{mu_weight object in grams}

\item{fleet_month}{vector of months (1-12) when sampling occurs}
}
\description{
Function to write an empirical weight at age SS file (wt_at_age.ss) from Atlantis output
}
\author{
Christine Stawitz
}
